Add ExclusiveBetween extension and test cases for classes
